README.md 	added install notes
gameslistEditor.py
	added support for kidgame, favorite, and hidden xml attributes
	scraper no longer works, will need to get it fixed or remove it

getIp.sh
	print ip address to shell script

makeBoxArtSymLinks.py
	fixed minor bug where emulationStation scraper would cause this
	to no longer worked.

startBrowserService.sh
	Script to start the browser service. I don't like to run the
	service all the time and I wanted to be able to start it with a
	controller so I've added this script to the config section of
	emulationStation.

+++ makeBoxArtSymLinks.py
@@ -27,5 +27,6 @@ for img in glob.glob(root + '/RetroPie/roms/*/downloaded_images_large/*'):
         continue
 
     print 'ln -s', fn
-    os.symlink(img, targetLinkName)
+    if not os.path.exists(targetLinkName):
+        os.symlink(img, targetLinkName)
